requirements/base.txt
appdirs==1.4.3
bugsnag==3.1.0
Django==1.11.1
django-admin-bootstrapped==2.5.7
django-bootstrap3==8.2.3
django-debug-toolbar==1.8
djangorestframework==3.6.3
factory-boy==2.8.1
fake-factory==9999.9.9
Faker==0.7.12
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
Unipath==1.1
WebOb==1.7.2
